Move tests to InteractionClients
