+++ lib/msal-browser/test/interaction_client/SilentCacheClient.spec.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+describe("SilentCacheClient", () => {
+    describe("acquireToken", () => {
+        it("successfully acquires the token from the cache", () => {
+
+        });
+    });
+});
